### 2021-06-25 #3411 @whedon
[PRE REVIEW]: rfasst: An R tool to estimate air pollution impacts on health and agriculture
**Submitting author:** @jonsampedro (<a href="http://orcid.org/0000-0002-2277-1530">Jon Sampedro</a>)
**Repository:** <a href="https://github.com/JGCRI/rfasst" target ="_blank">https://github.com/JGCRI/rfasst</a>
**Version:** v1.1
