### 2021-06-25 #3411 @kyleniemeyer
Hi @jonsampedro, thanks for your interest in JOSS!

Unfortunately, we have extremely limited editor availability to handle submissions right now, and so we need to put your submission on our waitlist untill an appropriate editor becomes available. Thanks for your patience!

Also, it looks like some references in your paper may be missing some DOIs—could you add those?
