### 2021-06-25 302d819
Add JOSS submitted badge and DOI to references
