+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={136},
   pages={105513},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.envint.2020.105513}
 }
 
 @article{markandya2018health,
@@ -16,7 +17,8 @@
   number={3},
   pages={e126--e133},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/s2542-5196(18)30029-9}
 }
 
 @article{calvin2019gcam,
@@ -37,7 +39,8 @@
   volume={231},
   pages={117538},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1002/essoar.10501661.1}
 }
 
 @article{emberson2018ozone,
@@ -47,7 +50,8 @@
   volume={100},
   pages={19--34},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.eja.2018.06.002}
 }
 
 @article{burnett2018global,
@@ -58,7 +62,8 @@
   number={38},
   pages={9592--9597},
   year={2018},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  DOI={10.1073/pnas.1803222115}
 }
 
 @article{van2018tm5,
@@ -69,14 +74,16 @@
   number={21},
   pages={16173--16211},
   year={2018},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  DOI={10.5194/acp-18-16173-2018}
 }
 
 @article{kyle2011gcam,
   title={GCAM 3.0 agriculture and land use: data sources and methods},
   author={Kyle, G Page and Luckow, Patrick and Calvin, Katherine V and Emanuel, William R and Nathan, Mayda and Zhou, Yuyu},
   year={2011},
-  institution={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)}
+  institution={Pacific Northwest National Lab.(PNNL), Richland, WA (United States)},
+  DOI={10.2172/1036082}
 }
 
 @article{van2009global,
@@ -87,7 +94,8 @@
   number={3},
   pages={604--618},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  DOI={10.1016/j.atmosenv.2008.10.033}
 }
 
 
@@ -99,5 +107,6 @@
   number={4},
   pages={1443--1475},
   year={2019},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  DOI={10.5194/gmd-12-1443-2019}
 }
