### 2021-06-25 #3411 @jonsampedro
Hello @kyleniemeyer, no problem, thanks for the quick response!

I have added the missing DOIs to the paper. Thanks!
