### 2021-11-01 #3820 @nanhung
@jonsampedro I can download the example GCAM database from Zenodo through `rpackageutils::download_unpack_zip()`, but have a problem conducting the following process. Can you figure out this issue?

![Screenshot 2021-11-01 131017](https://user-images.githubusercontent.com/25087463/139735289-70c38f1b-9c3e-4d1d-a913-a3dca97925c0.png)
