### 2021-11-02 #3820 @jonsampedro
@nanhung That error is related with the `rgcam` package (dependency), which is used to extract the relevant data from the GCAM database.  Is it your working directory (WD) equal to the "db_path"? `rgcam::localDBConn` does not work when the WD is set to be the same as the directory that holds GCAM database, so that may be the problem.
