### 2021-11-02 JGCRI/rfasst#58 @ibarraespinosa
```
options(timeout=2000)
#https://stackoverflow.com/questions/35282928/how-do-i-set-a-timeout-for-utilsdownload-file-in-r/35283374
setwd("rfasst/")
library(rfasst)
rpackageutils::download_unpack_zip(
  data_directory = "./inst/extdata",
  url = "https://zenodo.org/record/4763523/files/database_basexdb_5p3_release.zip?download=1")

setwd("..")

file.copy("rfasst/tests/testthat/inst/extdata/queries_rfasst.xml", ".")
download.file(
  url = "https://raw.githubusercontent.com/JGCRI/rfasst/main/tests/testthat/inst/extdata/queries_rfasst.xml",
  destfile = "queries_rfasst.xml"
)

file.remove("testJOSS.dat")

# To run a function from module 1:
m1_emissions_rescale(db_path =  "rfasst/inst/extdata/",
                     query_path="rfasst/inst/extdata/",
                     db_name = "database_basexdb_5p3_release",
                     prj_name = "testJOSS.dat",
                     scen_name = "Reference_gcam5p3_release",
                     queries ="queries_rfasst.xml",
                     saveOutput = F,
                     map = F)
```

Here only works with saveOutput = F and map = F

I haven't read al the documentation but it seems that the main output is testJOSS.dat. I will study yourpackage.
