### 2021-11-02 JGCRI/rfasst#58 @ibarraespinosa
When I use map = T I get the following:

```
Error in rmap::map(data = final_df_wide.map, shape = fasstSubset, folder = "output/maps/m1/maps_em",  :
  unused arguments (mapTitleOn = F, facetCols = 3, legendOutsideSingle = T, legendPosition = c("RIGHT", "bottom"), legendTextSizeO = 0.5, legendTitleSizeO = 0.7)

```
Also, should I need to create a directory output/maps/m1/?

Thanks
