### 2021-11-02 94ee969
Modifications for the maps to be consistent with the updates on rmap
