### 2021-11-02 JGCRI/rfasst#58 @jonsampedro
Hello @ibarraespinosa . Just for clarification, the main outputs from the package are the csv files and the maps; the `testJOSS.dat` is a data file to store the relevant information from the GCAM DB and avoid the need for querying the DB everytime a function is run.

Regarding the  saveOutput = F and map = F, as you say, you would need to create that directory for writing the outputs (`output/maps/m1/` ), but if you clone the repo, the folders are already there. Considering the number of diverse results that the tool generates, I organized the outputs so that each module produces the files in an individual folder.

I have also realized that there was an update in the downstrem **rmap** software, so if you activated map=T, there were some errors (as you indicate in the comment). I have made some modifications and the maps can be now generated with no issues, so you may need to the pull the changes, assuming you are testing the main branch, right? I could make a new release, but I thought that it would be more efficient making it after including all your comments... whatever works best!
