### 2021-11-02 #3820 @jonsampedro
Hello @nanhung . During the review process I have realized that there has been an update in the downstrem `rmap` software, so if you activated map=T, there were some errors. I have made some modifications and the maps can be now generated with no issues, so you may need to the pull the changes from the main branch. I could make a new release, but I thought that it would be more efficient making it after including all your comments. Sorry for the inconvenience!
