+++ .github/workflows/rcmd.yml
@@ -25,6 +25,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@master
+      - uses: r-lib/actions/setup-pandoc@v1
       
       - name: Brew and macOS config
         if: runner.os == 'macOS'
