### 2021-11-03 a5eb26c
Add Rshell to test_coverage
