+++ .github/workflows/test_coverage.yml
@@ -17,17 +17,18 @@ jobs:
           r-version: ${{ matrix.r }}
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
-      - uses: r-lib/actions/setup-pandoc@master
-      
+            
       - name: Install dependencies
         env:
           R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
           GITHUB_PAT: ${{ secrets.GH_PAT_RFASST}}
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr'))" -e "remotes::install_deps(dependencies = TRUE)"
+      
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(error_on = 'error')"
+      
       - name: Test coverage
-        if: matrix.r == '3.6'
         run: |
           remotes::install_cran("covr")
           covr::codecov(token = "${{secrets.CODECOV_TOKEN}}")
+        shell: Rscript {0}
