### 2021-11-08 #3820 @nanhung
@jonsampedro Thanks for your solution. However, I still can not solve the issue. The error occurs even I follow the tutorial in `rgcam`. Did I miss anything?

![Screenshot 2021-11-08 140237](https://user-images.githubusercontent.com/25087463/140826041-f3b3d184-418d-4600-a2a3-60f24979bf6f.png)
