### 2021-11-09 #3820 @jonsampedro
@nanhung I have replicated the tutorial example and the connection worked fine for me (see below).
![image](https://user-images.githubusercontent.com/78936824/140843363-fe9010b6-3b71-4979-b83a-05464c1f2bae.png)

Do you meet the system and package requirements to run the package? These are listed in the DESCRIPTION file at the `rgcam` repo: https://github.com/JGCRI/rgcam/blob/master/DESCRIPTION

I have been checking the rgcam issues site (https://github.com/JGCRI/rgcam/issues), and I cannot find a similar error listed there, but I will try to solve it and get you back. Thanks!
