### 2021-11-09 #3820 @nanhung
@jonsampedro Thanks! I got the problem solved by switching to Linux. There are some problems when running in Windows and even WSL (I guess it's from `rgcam` package). You might need to check if you can run the package under these OS. I will further test the example modules.
