### 2021-11-09 #3820 @nanhung
I found that the default setting is assumed that the user had `output/m1` and  `output/maps/m1/maps_em` directory to store the data for the m1 module. But it could be better if you add some code in your `m1_emissions_rescale` (and other modules) to create the directory if users didn't have this directory. For example,

```r
if (!dir.exists("output")) dir.create("output")
```
