### 2021-11-09 #3820 @jonsampedro
@nanhung yes I organized the outputs on different folders due to the large amount of diverse outputs that the tool can produce. If the repository is cloned the subdirectories will be already there, but I will add the suggested code to each of the modules to ensure the function still works even though there is no subfolder on the WD. Thanks for flagging this!
