### 2021-11-11 782e60a
Create the output subdirectories if they don't exist in the WD
