+++ R/m3_health.R
@@ -187,6 +187,13 @@ calc_daly_pm25<-function(){
 
 m3_get_mort_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,ssp="SSP2",saveOutput=T,map=F){
 
+  # Create the directories if they do not exist:
+  if (!dir.exists("output")) dir.create("output")
+  if (!dir.exists("output/m3")) dir.create("output/m3")
+  if (!dir.exists("output/maps")) dir.create("output/maps")
+  if (!dir.exists("output/maps/m3")) dir.create("output/maps/m3")
+  if (!dir.exists("output/maps/m3/maps_pm25_mort")) dir.create("output/maps/m3/maps_pm25_mort")
+
   # Ancillary Functions
   `%!in%` = Negate(`%in%`)
 
@@ -338,6 +345,13 @@ m3_get_mort_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries
 
 m3_get_mort_pm25_ecoloss<-function(db_path,query_path,db_name,prj_name,scen_name,ssp="SSP2",queries,saveOutput=T,map=F){
 
+  # Create the directories if they do not exist:
+  if (!dir.exists("output")) dir.create("output")
+  if (!dir.exists("output/m3")) dir.create("output/m3")
+  if (!dir.exists("output/maps")) dir.create("output/maps")
+  if (!dir.exists("output/maps/m3")) dir.create("output/maps/m3")
+  if (!dir.exists("output/maps/m3/maps_pm25_mort_ecoloss")) dir.create("output/maps/m3/maps_pm25_mort_ecoloss")
+
   # Ancillary Functions
   `%!in%` = Negate(`%in%`)
 
@@ -456,6 +470,13 @@ m3_get_mort_pm25_ecoloss<-function(db_path,query_path,db_name,prj_name,scen_name
 
 m3_get_yll_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,ssp="SSP2",saveOutput=T,map=F){
 
+  # Create the directories if they do not exist:
+  if (!dir.exists("output")) dir.create("output")
+  if (!dir.exists("output/m3")) dir.create("output/m3")
+  if (!dir.exists("output/maps")) dir.create("output/maps")
+  if (!dir.exists("output/maps/m3")) dir.create("output/maps/m3")
+  if (!dir.exists("output/maps/m3/maps_pm25_yll")) dir.create("output/maps/m3/maps_pm25_yll")
+
   # Ancillary Functions
   `%!in%` = Negate(`%in%`)
 
@@ -570,6 +591,13 @@ m3_get_yll_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,
 
 m3_get_yll_pm25_ecoloss<-function(db_path,query_path,db_name,prj_name,scen_name,queries,ssp="SSP2",saveOutput=T,map=F){
 
+  # Create the directories if they do not exist:
+  if (!dir.exists("output")) dir.create("output")
+  if (!dir.exists("output/m3")) dir.create("output/m3")
+  if (!dir.exists("output/maps")) dir.create("output/maps")
+  if (!dir.exists("output/maps/m3")) dir.create("output/maps/m3")
+  if (!dir.exists("output/maps/m3/maps_pm25_yll_ecoloss")) dir.create("output/maps/m3/maps_pm25_yll_ecoloss")
+
   # Ancillary Functions
   `%!in%` = Negate(`%in%`)
 
@@ -682,6 +710,13 @@ m3_get_yll_pm25_ecoloss<-function(db_path,query_path,db_name,prj_name,scen_name,
 #' @export
 
[…]
