+++ paper/paper.md
@@ -42,7 +42,25 @@ The structure of the `rfasst` package is summarized in Figure 1.
 
 ![Structure of the `rfasst` package](figure_rfasst.png)
 
-`rfasst` can be accessed via the web at the public domain https://github.com/JGCRI/rfasst. We provide an R vignette step-by-step tutorial for users to get started with `rfasst` which is accessible here: [Tutorial](https://jgcri.github.io/rfasst/).
+`rfasst` can be accessed via the web at the public domain https://github.com/JGCRI/rfasst. The following code is a simplified example that shows how to run the package. In addition, we provide an R vignette step-by-step tutorial for users to get started with `rfasst` which is accessible here: [Tutorial](https://jgcri.github.io/rfasst/).
+
+```r
+install.packages("devtools")
+library(devtools)
+devtools::install_github("JGCRI/rfasst")
+library(rfasst)
+
+db_path<-"path/to/my-gcam-database"
+query_path<-"path/to/queries_rfasst.xml"
+db_name<-"my-gcam-database"
+prj_name<-"my-project.dat" # This is a data file that is automatically generated to store the relevant information extracted from the GCAM DB.
+scen_name<-"name of the GCAM scenario" # The scenario name should be exactly equal to the name on the GCAM database.
+queries<-queries_rfasst.xml # The package includes a default query file that includes all the queries required in all the functions of the package, but it can be edited or changed by the user.
+
+# Using the m1_emissions_rescale function as a representative example. All the functions included in the packages are listed in the documentation vignettes.
+m1_emissions_rescale(db_path,query_path,db_name,prj_name,scen_name,queries,saveOutput=T, map=T)  
+
+```
 
 
 # Statement of need
@@ -61,7 +79,15 @@ and meteorological information required to translate the greenhouse gas and air 
 This transformation from emissions to concentration is addressed by full chemistry models or by simplified air quality emulators, such as TM5-FASST [@van2018tm5].
 These concentration levels are the inputs for the exposure-response functions that are normally used to calculate adverse human-health and agricultural effects associated with exposure to $PM_{2.5}$ and $O_{3}$.  
 
-Therefore, the combined use of these models, which is the essence of `rfasst`, is a powerful methodology to estimate a consistent range of health and agricultural damages and the co-benefits associated with different strategies or climate policies.
+Therefore, the combined use of these models, which is the essence of `rfasst`, is a powerful methodology to estimate a consistent range of health and agricultural damages and the co-benefits associated with different strategies or climate policies, that complements existing models and methods. 
+Historically, the assessment of air pollution and the subsequent impacts has been developed with global chemical transport models (CTMs) such as GEOS-Chem [@bey2001global] or the The Community Multiscale Air Quality Modeling System (CMAQ; @us_epa_CMAQ). 
+These models combine atmospheric, meteorological, and land-use data with anthropogenic emissions, which can be extracted from global emissions inventories, such as  The Emissions Database for Global Atmospheric Research (EDGAR; @crippa2016forty) or The Community Emissions Data System (CEDS; @hoesly2018historical), 
+or generated by additional tools, namely SMOKE [@smoke] or EmissV [@schuch2018emissv]. 
+While these CTMs provide detailed outputs, they are computationally expensive, so they have limited application for the assessment of several different narratives. 
+In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2020health], 
+has enabled to estimate air pollution and the associated impacts for global or regional changes in emission levels. 
[…]
