+++ paper/paper.bib
@@ -110,3 +110,104 @@
   publisher={Copernicus GmbH},
   DOI={10.5194/gmd-12-1443-2019}
 }
+
+@software{us_epa_CMAQ,
+  author       = {US EPA Office of Research and Development},
+  title        = {CMAQ},
+  month        = aug,
+  year         = 2021,
+  note         = {{For up-to-date documentation, source code, and 
+                   sample run scripts, please clone or download the
+                   CMAQ git repository available through GitHub:
+                   https://github.com/USEPA/CMAQ}},
+  publisher    = {Zenodo},
+  version      = {5.3.3},
+  doi          = {10.5281/zenodo.5213949},
+  url          = {https://doi.org/10.5281/zenodo.5213949}
+}
+
+@article{bey2001global,
+  title={Global modeling of tropospheric chemistry with assimilated meteorology: Model description and evaluation},
+  author={Bey, Isabelle and Jacob, Daniel J and Yantosca, Robert M and Logan, Jennifer A and Field, Brendan D and Fiore, Arlene M and Li, Qinbin and Liu, Honguy Y and Mickley, Loretta J and Schultz, Martin G},
+  journal={Journal of Geophysical Research: Atmospheres},
+  volume={106},
+  number={D19},
+  pages={23073--23095},
+  year={2001},
+  publisher={Wiley Online Library}
+}
+
+
+@article{hoesly2018historical,
+  title={Historical (1750--2014) anthropogenic emissions of reactive gases and aerosols from the Community Emissions Data System (CEDS)},
+  author={Hoesly, Rachel M and Smith, Steven J and Feng, Leyang and Klimont, Zbigniew and Janssens-Maenhout, Greet and Pitkanen, Tyler and Seibert, Jonathan J and Vu, Linh and Andres, Robert J and Bolt, Ryan M and others},
+  journal={Geoscientific Model Development},
+  volume={11},
+  number={1},
+  pages={369--408},
+  year={2018},
+  publisher={Copernicus GmbH}
+}
+
+@article{crippa2016forty,
+  title={Forty years of improvements in European air quality: regional policy-industry interactions with global impacts},
+  author={Crippa, Monica and Janssens-Maenhout, Greet and Dentener, Frank and Guizzardi, Diego and Sindelarova, Katerina and Muntean, Marilena and Van Dingenen, Rita and Granier, Claire},
+  journal={Atmospheric Chemistry and Physics},
+  volume={16},
+  number={6},
+  pages={3825--3841},
+  year={2016},
+  publisher={Copernicus GmbH}
+}
+
+
+@software{smoke,
+  author       = {B.H. Baek and
+                  Catherine Seppanen},
+  title        = {{CEMPD/SMOKE: SMOKE v4.8.1 Public Release (January 
+                   29, 2021)}},
+  month        = jan,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {SMOKEv481\_Jan2021},
+  doi          = {10.5281/zenodo.4480334},
+  url          = {https://doi.org/10.5281/zenodo.4480334}
+}
+
+
+@article{schuch2018emissv,
+  title={EmissV: an R package to create vehicular and other emissions for air quality models},
+  author={Schuch, Daniel and Ibarra-Espinosa, Sergio and de Freitas, Edmilson Dias},
+  journal={Journal of Open Source Software},
+  volume={3},
+  number={30},
+  pages={662},
+  year={2018}
+}
+
+@article{sacks2018environmental,
+  title={The Environmental Benefits Mapping and Analysis Program--Community Edition (BenMAP--CE): A tool to estimate the health and economic benefits of reducing air pollution},
[…]
