+++ paper/paper.md
@@ -53,12 +53,14 @@ library(rfasst)
 db_path<-"path/to/my-gcam-database"
 query_path<-"path/to/queries_rfasst.xml"
 db_name<-"my-gcam-database"
-prj_name<-"my-project.dat" # This is a data file that is automatically generated to store the relevant information extracted from the GCAM DB.
-scen_name<-"name of the GCAM scenario" # The scenario name should be exactly equal to the name on the GCAM database.
-queries<-queries_rfasst.xml # The package includes a default query file that includes all the queries required in all the functions of the package, but it can be edited or changed by the user.
+prj_name<-"my-project.dat"
+scen_name<-"name of the GCAM scenario"
+queries<-queries_rfasst.xml 
 
-# Using the m1_emissions_rescale function as a representative example. All the functions included in the packages are listed in the documentation vignettes.
-m1_emissions_rescale(db_path,query_path,db_name,prj_name,scen_name,queries,saveOutput=T, map=T)  
+# Using the m1_emissions_rescale function as a representative example. 
+# All the functions are listed in the documentation vignettes.
+m1_emissions_rescale(db_path,query_path,db_name,prj_name,
+		     scen_name,queries,saveOutput=T, map=T)  
 
 ```
 
@@ -80,11 +82,11 @@ This transformation from emissions to concentration is addressed by full chemist
 These concentration levels are the inputs for the exposure-response functions that are normally used to calculate adverse human-health and agricultural effects associated with exposure to $PM_{2.5}$ and $O_{3}$.  
 
 Therefore, the combined use of these models, which is the essence of `rfasst`, is a powerful methodology to estimate a consistent range of health and agricultural damages and the co-benefits associated with different strategies or climate policies, that complements existing models and methods. 
-Historically, the assessment of air pollution and the subsequent impacts has been developed with global chemical transport models (CTMs) such as GEOS-Chem [@bey2001global] or the The Community Multiscale Air Quality Modeling System (CMAQ; @us_epa_CMAQ). 
+Historically, the assessment of air pollution and the subsequent impacts has been developed with global chemical transport models (CTMs) such as GEOS-Chem [@bey2001global] or the The Community Multiscale Air Quality Modeling System (CMAQ; @byun1999science). 
 These models combine atmospheric, meteorological, and land-use data with anthropogenic emissions, which can be extracted from global emissions inventories, such as  The Emissions Database for Global Atmospheric Research (EDGAR; @crippa2016forty) or The Community Emissions Data System (CEDS; @hoesly2018historical), 
 or generated by additional tools, namely SMOKE [@smoke] or EmissV [@schuch2018emissv]. 
 While these CTMs provide detailed outputs, they are computationally expensive, so they have limited application for the assessment of several different narratives. 
-In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2020health], 
+In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2018airq+], 
 has enabled to estimate air pollution and the associated impacts for global or regional changes in emission levels. 
 In this framework, we believe that `rfasst` is a significant contribution to the community as it directly incorporates air pollution and its impacts into the integrated assessment of alternative “what-if” type global scenario analysis.
 
