+++ paper/paper.bib
@@ -111,19 +111,11 @@
   DOI={10.5194/gmd-12-1443-2019}
 }
 
-@software{us_epa_CMAQ,
-  author       = {US EPA Office of Research and Development},
-  title        = {CMAQ},
-  month        = aug,
-  year         = 2021,
-  note         = {{For up-to-date documentation, source code, and 
-                   sample run scripts, please clone or download the
-                   CMAQ git repository available through GitHub:
-                   https://github.com/USEPA/CMAQ}},
-  publisher    = {Zenodo},
-  version      = {5.3.3},
-  doi          = {10.5281/zenodo.5213949},
-  url          = {https://doi.org/10.5281/zenodo.5213949}
+@article{byun1999science,
+  title={Science algorithms of the EPA Models-3 community multiscale air quality (CMAQ) modeling system},
+  author={Byun, D},
+  journal={EPA/600/R-99/030},
+  year={1999}
 }
 
 @article{bey2001global,
@@ -205,9 +197,8 @@
   publisher={Elsevier}
 }
 
-@techreport{world2020health,
-  title={Health impact assessment of air pollution: AirQ+ life table manual, December 2020},
-  author={World Health Organization and others},
-  year={2020},
-  institution={World Health Organization. Regional Office for Europe}
+@misc{world2018airq+,
+  title={AirQ+: Software tool for health risk assessment of air pollution},
+  author={World Health Organization},
+  year={2018}
 }
