+++ paper/paper.bib
@@ -197,7 +197,7 @@
   publisher={Elsevier}
 }
 
-@misc{world2018airq+,
+@techreport{world2018airq,
   title={AirQ+: Software tool for health risk assessment of air pollution},
   author={World Health Organization},
   year={2018}
