+++ paper/paper.md
@@ -86,7 +86,7 @@ Historically, the assessment of air pollution and the subsequent impacts has bee
 These models combine atmospheric, meteorological, and land-use data with anthropogenic emissions, which can be extracted from global emissions inventories, such as  The Emissions Database for Global Atmospheric Research (EDGAR; @crippa2016forty) or The Community Emissions Data System (CEDS; @hoesly2018historical), 
 or generated by additional tools, namely SMOKE [@smoke] or EmissV [@schuch2018emissv]. 
 While these CTMs provide detailed outputs, they are computationally expensive, so they have limited application for the assessment of several different narratives. 
-In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2018airq+], 
+In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2018airq], 
 has enabled to estimate air pollution and the associated impacts for global or regional changes in emission levels. 
 In this framework, we believe that `rfasst` is a significant contribution to the community as it directly incorporates air pollution and its impacts into the integrated assessment of alternative “what-if” type global scenario analysis.
 
