+++ paper/paper.bib
@@ -199,6 +199,6 @@
 
 @techreport{world2018airq,
   title={AirQ+: Software tool for health risk assessment of air pollution},
-  author={World Health Organization},
+  institution={World Health Organization},
   year={2018}
 }
