+++ paper/paper.bib
@@ -198,6 +198,7 @@
 }
 
 @techreport{world2018airq,
+  author={WHO}
   title={AirQ+: Software tool for health risk assessment of air pollution},
   institution={World Health Organization},
   year={2018}
