+++ paper/paper.md
@@ -85,10 +85,10 @@ Therefore, the combined use of these models, which is the essence of `rfasst`, i
 Historically, the assessment of air pollution and the subsequent impacts has been developed with global chemical transport models (CTMs) such as GEOS-Chem [@bey2001global] or the The Community Multiscale Air Quality Modeling System (CMAQ; @byun1999science). 
 These models combine atmospheric, meteorological, and land-use data with anthropogenic emissions, which can be extracted from global emissions inventories, such as  The Emissions Database for Global Atmospheric Research (EDGAR; @crippa2016forty) or The Community Emissions Data System (CEDS; @hoesly2018historical), 
 or generated by additional tools, namely SMOKE [@smoke] or EmissV [@schuch2018emissv]. 
-While these CTMs provide detailed outputs, they are computationally expensive, so they have limited application for the assessment of several different narratives. 
-In order to provide more rapid and flexible (but less detailed) analysis, the development of different air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2018airq], 
-has enabled to estimate air pollution and the associated impacts for global or regional changes in emission levels. 
-In this framework, we believe that `rfasst` is a significant contribution to the community as it directly incorporates air pollution and its impacts into the integrated assessment of alternative “what-if” type global scenario analysis.
+While these CTMs provide detailed outputs, they are also computationally expensive, so they have limited application for either rapid assessment or evaluation of multiple scenarios.
+A number of air quality tools and emulators, such as TM5-FASST [@van2018tm5], BenMap [@sacks2018environmental], SHERPA [@clappier2015new] or AirQ+ [@world2018airq], 
+have been developed to allow faster estimation of air pollution levels and associated impacts for global or regional changes in emission levels. 
+`rfasst` contributes to this set of tools by making the already widely used TM5-FASST tool available as an easily accessible `R` package that can be readily used for air pollution and impact assessment of alternative “what-if” global scenarios.
 
 Prior to the development of this package, we have used GCAM and TM5-FASST to analyze these co-benefits in different studies. We showed that health co-benefits attributable to air pollution are larger than mitigation costs 
 for different technological scenarios consistent with the 2°C target of the Paris Agreement [@sampedro_2020a]. 
