### 2021-11-11 c4d143a
Add DOI to new references
