+++ paper/paper.bib
@@ -111,11 +111,16 @@
   DOI={10.5194/gmd-12-1443-2019}
 }
 
-@article{byun1999science,
-  title={Science algorithms of the EPA Models-3 community multiscale air quality (CMAQ) modeling system},
-  author={Byun, D},
-  journal={EPA/600/R-99/030},
-  year={1999}
+@article{appel2021community,
+  title={The Community Multiscale Air Quality (CMAQ) model versions 5.3 and 5.3. 1: system updates and evaluation},
+  author={Appel, K Wyat and Bash, Jesse O and Fahey, Kathleen M and Foley, Kristen M and Gilliam, Robert C and Hogrefe, Christian and Hutzell, William T and Kang, Daiwen and Mathur, Rohit and Murphy, Benjamin N and others},
+  journal={Geoscientific Model Development},
+  volume={14},
+  number={5},
+  pages={2867--2897},
+  year={2021},
+  publisher={Copernicus GmbH}
+  DOI={10.5194/gmd-14-2867-2021}
 }
 
 @article{bey2001global,
@@ -127,6 +132,7 @@
   pages={23073--23095},
   year={2001},
   publisher={Wiley Online Library}
+  DOI={10.1029/2001JD000807}
 }
 
 
@@ -139,6 +145,7 @@
   pages={369--408},
   year={2018},
   publisher={Copernicus GmbH}
+  DOI={10.5194/gmd-11-369-2018}
 }
 
 @article{crippa2016forty,
@@ -150,6 +157,7 @@
   pages={3825--3841},
   year={2016},
   publisher={Copernicus GmbH}
+  DOI={10.5194/acp-16-3825-2016}
 }
 
 
@@ -175,6 +183,7 @@
   number={30},
   pages={662},
   year={2018}
+  DOI={10.21105/joss.00662}
 }
 
 @article{sacks2018environmental,
@@ -185,6 +194,7 @@
   pages={118--129},
   year={2018},
   publisher={Elsevier}
+  DOI={10.1016/j.envsoft.2018.02.009}
 }
 
 @article{clappier2015new,
@@ -195,11 +205,13 @@
   pages={66--74},
   year={2015},
   publisher={Elsevier}
+  DOI={10.1016/j.envsoft.2015.09.007}
 }
 
-@techreport{world2018airq,
-  author={WHO}
+
+@Manual{world2018airq,
   title={AirQ+: Software tool for health risk assessment of air pollution},
-  institution={World Health Organization},
-  year={2018}
+  author = {WHO},
+  year = {2018},
+  url = {https://www.euro.who.int/en/health-topics/environment-and-health/air-quality/activities/airq-software-tool-for-health-risk-assessment-of-air-pollution},
 }
