+++ paper/paper.bib
@@ -119,7 +119,7 @@
   number={5},
   pages={2867--2897},
   year={2021},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
   DOI={10.5194/gmd-14-2867-2021}
 }
 
@@ -131,7 +131,7 @@
   number={D19},
   pages={23073--23095},
   year={2001},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   DOI={10.1029/2001JD000807}
 }
 
@@ -144,7 +144,7 @@
   number={1},
   pages={369--408},
   year={2018},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
   DOI={10.5194/gmd-11-369-2018}
 }
 
@@ -156,7 +156,7 @@
   number={6},
   pages={3825--3841},
   year={2016},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
   DOI={10.5194/acp-16-3825-2016}
 }
 
@@ -182,7 +182,7 @@
   volume={3},
   number={30},
   pages={662},
-  year={2018}
+  year={2018},
   DOI={10.21105/joss.00662}
 }
 
@@ -193,7 +193,7 @@
   volume={104},
   pages={118--129},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
   DOI={10.1016/j.envsoft.2018.02.009}
 }
 
@@ -204,7 +204,7 @@
   volume={74},
   pages={66--74},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
   DOI={10.1016/j.envsoft.2015.09.007}
 }
 
@@ -213,5 +213,5 @@
   title={AirQ+: Software tool for health risk assessment of air pollution},
   author = {WHO},
   year = {2018},
-  url = {https://www.euro.who.int/en/health-topics/environment-and-health/air-quality/activities/airq-software-tool-for-health-risk-assessment-of-air-pollution},
+  url = {https://www.euro.who.int/en/health-topics/environment-and-health/air-quality/activities/airq-software-tool-for-health-risk-assessment-of-air-pollution}
 }
