### 2021-11-11 #3820 @jonsampedro
Hello, I just wanted to let you know I have pushed the following changes to the main branch:

A suggested by @nanhung  , I have edited all the functions in the package, so that they create the corresponding subdirectory if it does not exist, instead of giving an error.

Following the comments by @ibarraespinosa , I have modified the main text and it now includes:
- A code example to show how the package works.
- A new framing of the tool, comparing the package with additional models and tools that have been extensively applied in scientific literature.

Hope these changes look Ok to you, and happy to make any additional changes and clarifications.
