### 2021-11-17 #3820 @nanhung
Thanks @jonsampedro ! I am still working on the module 1 and checking the details.
I suggest that you put more explanation of the output. For instance, where can find the full name of the abbreviation? Waht is the unit of the value?
You can add the description in `m1_emissions_rescale()` function.
