### 2021-11-19 39bdc36
Add sources and units to function documentation
