### 2021-11-19 #3820 @jonsampedro
@nanhung I have updated the function documentation (in the four modules) including all the units and the sources for:

-  Full names of the regions and list of the countries that form each region
-  List of commodities that form each crop category
