+++ vignettes/Module3_health.Rmd
@@ -100,7 +100,7 @@ The complete list of outputs generated by the suite of functions that form this 
 * `O3_MORT_ECOLOSS_[scenario]_[year].csv`
 * `O3_YLL_ECOLOSS__[scenario]_[year].csv`
 
-As in Module 2, for all these functions, the package allows to produce different figures and/or animations, generated using the [rmap](https://github.com/JGCRI/rmap) package documented in the following [page](jgcri.github.io/rmap/). To generate these maps, the user needs to include the `map=T` parameter, and they will be generated and stored in the corresponding output sub-directory. As an example for this module, the following map shows the premature mortalities attributable to PM2.5 exposure by cause in 2050.
+As in Module 2, for all these functions, the package allows to produce different figures and/or animations, generated using the [rmap](https://github.com/JGCRI/rmap) package documented in the following [page](https://jgcri.github.io/rmap/). To generate these maps, the user needs to include the `map=T` parameter, and they will be generated and stored in the corresponding output sub-directory. As an example for this module, the following map shows the premature mortalities attributable to PM2.5 exposure by cause in 2050.
 
 <!-------------------------->
 <!-------------------------->
