### 2021-12-16 JGCRI/rfasst#58 @ibarraespinosa
I have updated the package and now everything is running smoothly. Including combinations true and false for maps and output. Great!
Comment on the figures. The maps are using the same legend for different pollutants. I think it is better to produce one legend for each pollutan
![map_param_2080_PRETTY](https://user-images.githubusercontent.com/27447280/146406675-257c5c8f-78c2-4822-8b87-e0332fc975c0.png)
t.
