### 2021-12-16 JGCRI/rfasst#58 @ibarraespinosa
I have been following your tutorial and all is goof. map = T takes lots of time after the last image, idkw, but it looks good anyway. I will be running al code. In the meantime, I strongly suggest you to adopt a styling code. There are good examples, such as the guide of ropensci, tidyverse or the guide of google to write R code.
