### 2021-12-16 JGCRI/rfasst#58 @ibarraespinosa
i suggest you to add a logical argument to generate annimations
