### 2021-12-17 JGCRI/rfasst#58 @jonsampedro
Hello @ibarraespinosa . Thanks for your comments, I will make the suggested changes in the following days and let you know.

Regarding the styling code, I have tried to follow the Google's R Style Guide (https://google.github.io/styleguide/Rguide.html), so that the code includes left-hand assignments and explicitly qualifies namespaces. For the next version, I will add the explicit returns to my functions, and please let me know if you have any additional suggestions.

For the individualized legends and animations, some of these features are defined in the underlying `rmap` software, but we will work to implement your suggestions.

Thanks!
