### 2021-12-17 JGCRI/rfasst#58 @ibarraespinosa
That is good. I see that the software has improved and the paper as well.
Regarding the style of code, I strongly suggest add spaces and new lines
after commas.

In my opinion, your article and software is good enough to be accepted.

however, a strong suggestion is to include more guidelines to new users so
that they can use your software. For instance, I would like to try your
software and make some experiments for Brazil, but it can be a bit
overwhelming. Anyways, congratulations.

Em sex., 17 de dez. de 2021 às 19:45, jonsampedro ***@***.***>
escreveu:

> Hello @ibarraespinosa <https://github.com/ibarraespinosa> . Thanks for
> your comments, I will make the suggested changes in the following days and
> let you know.
>
> Regarding the styling code, I have tried to follow the Google's R Style
> Guide (https://google.github.io/styleguide/Rguide.html), so that the code
> includes left-hand assignments and explicitly qualifies namespaces. For the
> next version, I will add the explicit returns to my functions, and please
> let me know if you have any additional suggestions.
>
> For the individualized legends and animations, some of these features are
> defined in the underlying rmap software, but we will work to implement
> your suggestions.
>
> Thanks!
>
