### 2021-12-21 #3820 @ibarraespinosa
hello @dhhagan, I think the manuscript and software are good enough to be accepted. I have tested all the functions and the @jonsampedro  has incorporated many of my suggestions. Then I recommend acceptance of rfasst for publication at JOSS.

Merry Christmas!  🎅 🎄 🎶
