### 2021-12-21 JGCRI/rfasst#58 @jonsampedro
@ibarraespinosa, thank you for very constructive review. I will continue updating the software in line with your comments and suggestions. Happy Holidays!
