### 2021-12-22 64c0ab9
Add logical argument for animations, individual pollutants and additional styling changes
