+++ R/m3_health.R
@@ -41,7 +41,8 @@ calc_daly_tot<-function(){
     tidyr::spread(measure_name,val) %>%
     dplyr::rename(country=location_name) %>%
     gcamdata::left_join_error_no_match(country_iso, by="country") %>%
-    gcamdata::left_join_error_no_match(fasst_reg %>% dplyr::rename(iso3=subRegionAlt),
+    gcamdata::left_join_error_no_match(fasst_reg %>%
+                                         dplyr::rename(iso3=subRegionAlt),
                                        by="iso3") %>%
     dplyr::group_by(fasst_region,year,cause_name) %>%
     dplyr::summarise(`DALYs (Disability-Adjusted Life Years)`=sum(`DALYs (Disability-Adjusted Life Years)`),
@@ -62,7 +63,8 @@ calc_daly_tot<-function(){
     tidyr::spread(measure_name,val) %>%
     dplyr::rename(country=location_name) %>%
     gcamdata::left_join_error_no_match(country_iso, by="country") %>%
-    gcamdata::left_join_error_no_match(fasst_reg %>% dplyr::rename(iso3=subRegionAlt),
+    gcamdata::left_join_error_no_match(fasst_reg %>%
+                                         dplyr::rename(iso3=subRegionAlt),
                                        by="iso3") %>%
     dplyr::group_by(fasst_region,year,cause_name) %>%
     dplyr::summarise(`DALYs (Disability-Adjusted Life Years)`=sum(`DALYs (Disability-Adjusted Life Years)`),
@@ -103,7 +105,8 @@ calc_daly_o3<-function(){
     tidyr::spread(measure_name,val) %>%
     dplyr::rename(country=location_name) %>%
     gcamdata::left_join_error_no_match(country_iso, by="country") %>%
-    gcamdata::left_join_error_no_match(fasst_reg %>% dplyr::rename(iso3=subRegionAlt),
+    gcamdata::left_join_error_no_match(fasst_reg %>%
+                                         dplyr::rename(iso3=subRegionAlt),
                                        by="iso3") %>%
     dplyr::group_by(fasst_region,year,cause_name) %>%
     dplyr::summarise(`DALYs (Disability-Adjusted Life Years)`=sum(`DALYs (Disability-Adjusted Life Years)`),
@@ -147,7 +150,8 @@ calc_daly_pm25<-function(){
     tidyr::spread(measure_name,val) %>%
     dplyr::rename(country=location_name) %>%
     gcamdata::left_join_error_no_match(country_iso, by="country") %>%
-    gcamdata::left_join_error_no_match(fasst_reg %>% dplyr::rename(iso3=subRegionAlt),
+    gcamdata::left_join_error_no_match(fasst_reg %>%
+                                         dplyr::rename(iso3=subRegionAlt),
                                        by="iso3") %>%
     dplyr::group_by(fasst_region,year,cause_name) %>%
     dplyr::summarise(`DALYs (Disability-Adjusted Life Years)`=sum(`DALYs (Disability-Adjusted Life Years)`),
@@ -182,10 +186,12 @@ calc_daly_pm25<-function(){
 #' @param saveOutput Writes the emission files.By default=T
 #' @param ssp Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2
 #' @param map Produce the maps. By default=F
+#' @param anim If set to T, produces multi-year animations. By default=T
 #' @importFrom magrittr %>%
 #' @export
 
-m3_get_mort_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,ssp="SSP2",saveOutput=T,map=F){
[…]
