+++ R/m2_conc.R
@@ -12,10 +12,12 @@
 #' @param queries Name of the GCAM query file. The file by default includes the queries required to run rfasst
 #' @param saveOutput Writes the files.By default=T
 #' @param map Produce the maps. By default=F
+#' @param anim If set to T, produces multi-year animations. By default=T
 #' @importFrom magrittr %>%
 #' @export
 
-m2_get_conc_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,saveOutput=T,map=F){
+m2_get_conc_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries,
+                           saveOutput=T,map=F, anim=T){
 
   # Create the directories if they do not exist:
   if (!dir.exists("output")) dir.create("output")
@@ -46,13 +48,13 @@ m2_get_conc_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries
   base_conc<-raw.base_conc %>%
     tidyr::gather(pollutant,value,-COUNTRY,-AREA_M2,-POP) %>%
     dplyr::mutate(units=dplyr::if_else(pollutant %in% c("O3","M6M","M3M"),"ppbv","ug/m3"),
-           year="base") %>%
+                  year="base") %>%
     dplyr::rename(region=COUNTRY)
 
   base_em<-raw.base_em %>%
     tidyr::gather(pollutant,value,-COUNTRY) %>%
     dplyr::mutate(units="kt",
-           year="base")
+                  year="base")
 
 
   # Then, we load all the SRC matrix for PM2.5, O3 and the specific measures: AOT, Mi and M6M
@@ -88,17 +90,19 @@ m2_get_conc_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries
     gcamdata::repeat_add_columns(tibble::tibble(base_year=all_years)) %>%
     dplyr::select(-year) %>%
     dplyr::rename(year=base_year,
-           region=COUNTRY) %>%
+                  region=COUNTRY) %>%
     dplyr::filter(region != "*TOTAL*") %>%
     dplyr::mutate(year=as.factor(as.character(year))) %>%
-    gcamdata::left_join_error_no_match(dplyr::bind_rows(em.list) %>% dplyr::mutate(year=as.factor(year)),by=c("region","year","pollutant")) %>%
+    gcamdata::left_join_error_no_match(dplyr::bind_rows(em.list) %>%
+                                         dplyr::mutate(year=as.factor(year)),
+                                       by=c("region","year","pollutant")) %>%
     dplyr::mutate(value_div=value.x)
 
   # Calculated the normalized CH4 HTAP change
   delta_em_ch4_htap<-delta_em %>%
     dplyr::filter(pollutant=="CH4") %>%
     dplyr::mutate(pollutant="CH4_HTAP",
-           value_div=ch4_htap_pert)
+                  value_div=ch4_htap_pert)
 
   delta_em<-delta_em %>%
     dplyr::bind_rows(delta_em_ch4_htap) %>%
@@ -107,7 +111,7 @@ m2_get_conc_pm25<-function(db_path,query_path,db_name,prj_name,scen_name,queries
     dplyr::mutate(delta_em=dplyr::if_else(pollutant=="PM25",0,delta_em)) %>%
     dplyr::arrange(region) %>%
     dplyr::mutate(delta_em = replace(delta_em, is.nan(delta_em), 0),
-           delta_em = replace(delta_em, !is.finite(delta_em), 0),) %>%
+                  delta_em = replace(delta_em, !is.finite(delta_em), 0),) %>%
     dplyr::mutate(region=gsub("AIR","Air",region),
            region=gsub("SHIP","Ship",region))  %>%
[…]
