+++ vignettes/Module1_emissions.Rmd
@@ -25,7 +25,8 @@ Outputs:
 
 The function produces re-scaled emissions of the main pollutants for each period and TM5-FASST region. If `saveOutput` is set to `TRUE`, the function writes the following csv files in the `output/m1` sub-directory: `[scenario]_[year].csv`
 
-In addition, by setting `map` to `TRUE`, the function generates air pollutant emission maps by year and specie, using the [rmap](https://github.com/JGCRI/rmap) package documented in the following [page](https://jgcri.github.io/rmap/).
+In addition, by setting `map` to `TRUE`, the function generates air pollutant emission maps by year and specie, using the [rmap](https://github.com/JGCRI/rmap) package documented in the following [page](https://jgcri.github.io/rmap/). The function also 
+generates animations and individual figures for each pollutant by modifying the `mapIndivPol`and `anim` parameters.
 
 The outputs from this module generated with the function `m1_emissions_rescale` are going to read by the set of functions in Module 2, which calculate  PM2.5 and O3 average concentration levels based on these emission sets.
 
@@ -49,7 +50,7 @@ library(magrittr)
  queries<-"Name of the query file" # (the package includes a default query file that includes all the queries required in every function in the package, "queries_rfasst.xml")
 
    #To write the re-scaled emissions (csv files) for all years into the output folder:
-     m1_emissions_rescale(db_path,query_path,db_name,prj_name,scen_name,queries,saveOutput=T, map=T)
+     m1_emissions_rescale(db_path,query_path,db_name,prj_name,scen_name,queries,saveOutput=T, map=T,mapIndivPol=F, anim=T )
 
    #To save as data frame emisions of main pollutants by TM5-FASST region in 2050:
 
