+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 2.7.3
-pkgdown: 1.6.1
+pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   Module1_emissions: Module1_emissions.html
   Module2_concentration: Module2_concentration.html
   Module3_health: Module3_health.html
   Module4_agriculture: Module4_agriculture.html
-last_built: 2021-11-19T19:13Z
+last_built: 2021-12-22T14:27Z
 
