### 2022-01-11 #3820 @dhhagan
@jonsampedro Can you do the following for me?

- [ ] Make a tagged release and archive.
- [ ] Report the version number and archive doi in this thread

After that is complete, I can move ahead with the acceptance process.
