### 2022-01-11 #3820 @jonsampedro
Sure @dhhagan, I have created a new release (rfasst-v1.2) , which is stored in zenodo with the following details:

**Zenodo site**: https://zenodo.org/record/5838361#.Yd2ym_7MJPY
**DOI**: 10.5281/zenodo.5838361
