### 2022-01-18 #3820 @jonsampedro
Hello @dhhagan , I have edited the archive title and authors to match the paper. See the updated archive here: https://zenodo.org/record/5838361#.YebL0P7MKUk

Note that, for clarity, I have also edited the corresponding release in the GitHub repository (https://github.com/JGCRI/rfasst/releases/tag/rfasst-v1.2)

Hope everything is OK now, thanks!
